Use the hash of a node's key as the initialization vector for encrypting
Every object in the Cryptosphere is encrypted with a unique key. This
means the initialization vector is irrelevant as each key is by design
only used to encrypt one body of plaintext. One option would be to use
all zeroes, however for some added obfuscation the SHA256 of the
original key is computed and used as the IV.