Rip off the OpenSSL band-aid
Best line of this commit:
-require 'openssl'

This commit completely removes the OpenSSL dependency, replacing it with a new
initial head system based on Cryptosphere::WriteHeads and
Cryptosphere::ReadHeads for storing pointers to URIs in the system.

Cryptosphere::Identity has been deleted for now with no suitable replacement.

There should no longer be any shitty crypto in the Cryptosphere. It should be
gone now. Please go nuts looking for problems ;)