New logo (with new hidden message)
This is the third Cryptosphere logo puzzle. There is a new hidden message
encrypted in the binary numbers that line the logo image's background.

The previous logo's message was encrypted with a Vigenère cipher. This one uses
a more difficult cipher.

For more information on the logo ciphers, see:

https://github.com/cryptosphere/cryptosphere/wiki/Logo-Ciphers