We do not yet implement z-base-32
I appear to have misunderstood the exact nature of Zooko's Base32, which is
well-documented here:

http://philzimmermann.com/docs/human-oriented-base-32-encoding.txt

There are a number of other permutations that have been made to Base32 which are
not correctly captured in the Cryptosphere's present implementation. Among other
things characters which are easily confused have been removed.

The specific encoding specification should be considered TBD. In the meantime we
can at least correct the oversight and note that we are using a semi-proprietary
version of Base32 (although not really, this style is supported by things like
Google Authenticator)