Remove use of Blake2b hash keys
The previous schemes which passed short keys to Blake2b instead of
personalization schemes break under RbNaCl 2.0, which now imposes some security
constraints on the lengths of Blake2b keys.

For now, just get rid of the keys. The goal of this usage was to create
different hash families for these various usages. The value this adds is
debatable. For now, we'll get rid of them to get the code working again.